Changed The Terraforming
Removed the ability for tiles to self terraform (For gameplay reason and optomisation)
Now the player can spend resource to upgrade the terraform dome.
The resources are Carbon, Power and Water.
Upgrading the dome cost's carbon and power.
Currently, the starting resources are all the player has.
Next I will be adding in the ability to gain resoures.