Fixed TimeSkip / Added Weather Information to Ship
Fixed the ability to speed up time but had to remove the ability to pick a specific time to auto skip too. You now have to manually stop the skip.
Added the weather panel that used to be activated with the radio tower to the main ship.
Updated the UI on the weather panel to match the new UI.
Added indicator icons for when the game is paused, skipping time or playing